+++ mobile/src/main/java/com/github/shadowsocks/ProfileConfigFragment.kt
@@ -26,14 +26,14 @@ import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import android.os.UserManager
-import android.support.design.widget.Snackbar
-import android.support.v14.preference.SwitchPreference
-import android.support.v7.app.AlertDialog
-import android.support.v7.preference.Preference
-import android.support.v7.preference.PreferenceDataStore
-import android.support.v7.widget.Toolbar
 import android.view.MenuItem
-import com.github.shadowsocks.App.Companion.app
+import androidx.appcompat.app.AlertDialog
+import androidx.core.content.getSystemService
+import androidx.core.os.bundleOf
+import androidx.preference.Preference
+import androidx.preference.PreferenceDataStore
+import androidx.preference.SwitchPreference
+import com.github.shadowsocks.Core.app
 import com.github.shadowsocks.database.Profile
 import com.github.shadowsocks.database.ProfileManager
 import com.github.shadowsocks.plugin.PluginConfiguration
@@ -47,10 +47,10 @@ import com.github.shadowsocks.preference.PluginConfigurationDialogFragment
 import com.github.shadowsocks.utils.Action
 import com.github.shadowsocks.utils.DirectBoot
 import com.github.shadowsocks.utils.Key
-import com.takisoft.fix.support.v7.preference.EditTextPreference
-import com.takisoft.fix.support.v7.preference.PreferenceFragmentCompatDividers
+import com.takisoft.preferencex.EditTextPreference
+import com.takisoft.preferencex.PreferenceFragmentCompat
 
-class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenuItemClickListener,
+class ProfileConfigFragment : PreferenceFragmentCompat(),
         Preference.OnPreferenceChangeListener, OnPreferenceDataStoreChangeListener {
     companion object {
         private const val REQUEST_CODE_PLUGIN_CONFIGURE = 1
@@ -68,7 +68,7 @@ class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenu
         val activity = requireActivity()
         profileId = activity.intent.getLongExtra(Action.EXTRA_PROFILE_ID, -1L)
         addPreferencesFromResource(R.xml.pref_profile)
-        if (Build.VERSION.SDK_INT >= 25 && activity.getSystemService(UserManager::class.java).isDemoUser) {
+        if (Build.VERSION.SDK_INT >= 25 && activity.getSystemService<UserManager>()?.isDemoUser == true) {
             findPreference(Key.host).summary = "shadowsocks.example.org"
             findPreference(Key.remotePort).summary = "1337"
             findPreference(Key.password).summary = "\u2022".repeat(32)
@@ -93,12 +93,12 @@ class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenu
             pluginConfigure.isEnabled = newValue.isNotEmpty()
             pluginConfigure.text = pluginConfiguration.selectedOptions.toString()
             if (PluginManager.fetchPlugins()[newValue]?.trusted == false)
-                Snackbar.make(view!!, R.string.plugin_untrusted, Snackbar.LENGTH_LONG).show()
+                (activity as MainActivity).snackbar().setText(R.string.plugin_untrusted).show()
             true
         }
         pluginConfigure.onPreferenceChangeListener = this
         initPlugins()
-        receiver = app.listenForPackageChanges(false) { initPlugins() }
+        receiver = Core.listenForPackageChanges(false) { initPlugins() }
         DataStore.privateStore.registerChangeListener(this)
     }
 
@@ -116,12 +116,9 @@ class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenu
         pluginConfigure.text = pluginConfiguration.selectedOptions.toString()
     }
 
-    private fun showPluginEditor() {
-        val bundle = Bundle()
-        bundle.putString("key", Key.pluginConfigure)
-        bundle.putString(PluginConfigurationDialogFragment.PLUGIN_ID_FRAGMENT_TAG, pluginConfiguration.selected)
-        displayPreferenceDialog(PluginConfigurationDialogFragment(), Key.pluginConfigure, bundle)
-    }
+    private fun showPluginEditor() = displayPreferenceDialog(PluginConfigurationDialogFragment(), Key.pluginConfigure,
+            bundleOf(Pair("key", Key.pluginConfigure),
+                    Pair(PluginConfigurationDialogFragment.PLUGIN_ID_FRAGMENT_TAG, pluginConfiguration.selected)))
 
     fun saveAndExit() {
         val profile = ProfileManager.getProfile(profileId) ?: Profile()
@@ -146,7 +143,7 @@ class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenu
         DataStore.dirty = true
         true
     } catch (exc: IllegalArgumentException) {
-        Snackbar.make(view!!, exc.localizedMessage, Snackbar.LENGTH_LONG).show()
+        (activity as MainActivity).snackbar(exc.localizedMessage).show()
         false
     }
 
@@ -159,8 +156,7 @@ class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenu
             val intent = PluginManager.buildIntent(pluginConfiguration.selected, PluginContract.ACTION_CONFIGURE)
             if (intent.resolveActivity(requireContext().packageManager) == null) showPluginEditor() else
                 startActivityForResult(intent
-                        .putExtra(PluginContract.EXTRA_OPTIONS, pluginConfiguration.selectedOptions.toString())
-                        .putExtra(PluginContract.EXTRA_NIGHT_MODE, DataStore.nightMode),
+                        .putExtra(PluginContract.EXTRA_OPTIONS, pluginConfiguration.selectedOptions.toString()),
                         REQUEST_CODE_PLUGIN_CONFIGURE)
         } else super.onDisplayPreferenceDialog(preference)
     }
@@ -173,19 +169,18 @@ class ProfileConfigFragment : PreferenceFragmentCompatDividers(), Toolbar.OnMenu
                 onPreferenceChange(null, options)
             }
             PluginContract.RESULT_FALLBACK -> showPluginEditor()
-
         } else super.onActivityResult(requestCode, resultCode, data)
     }
 
-    override fun onMenuItemClick(item: MenuItem) = when (item.itemId) {
+    override fun onOptionsItemSelected(item: MenuItem?) = when (item?.itemId) {
         R.id.action_delete -> {
             val activity = requireActivity()
             AlertDialog.Builder(activity)
                     .setTitle(R.string.delete_confirm_prompt)
-                    .setPositiveButton(R.string.yes, { _, _ ->
+                    .setPositiveButton(R.string.yes) { _, _ ->
                         ProfileManager.delProfile(profileId)
                         activity.finish()
-                    })
+                    }
                     .setNegativeButton(R.string.no, null)
                     .create()
                     .show()